Implementing task #34, table now includes concentration ratios.
PyionData and PyionUnit classes were moved for organizational purposes.
